+++ src/main/java/edu/ie3/datamodel/io/source/sql/SqlTimeSeriesSource.java
@@ -7,9 +7,9 @@ package edu.ie3.datamodel.io.source.sql;
 
 import edu.ie3.datamodel.exceptions.SourceException;
 import edu.ie3.datamodel.io.connectors.SqlConnector;
-import edu.ie3.datamodel.io.csv.timeseries.IndividualTimeSeriesMetaInformation;
 import edu.ie3.datamodel.io.factory.timeseries.SimpleTimeBasedValueData;
 import edu.ie3.datamodel.io.factory.timeseries.TimeBasedSimpleValueFactory;
+import edu.ie3.datamodel.io.naming.timeseries.IndividualTimeSeriesMetaInformation;
 import edu.ie3.datamodel.io.source.TimeSeriesSource;
 import edu.ie3.datamodel.models.timeseries.individual.IndividualTimeSeries;
 import edu.ie3.datamodel.models.timeseries.individual.TimeBasedValue;
@@ -35,6 +35,36 @@ public class SqlTimeSeriesSource<V extends Value> extends SqlDataSource<TimeBase
 
   private final String queryTimeInterval;
   private final String queryTime;
+
+  /**
+   * Factory method to build a source from given meta information
+   *
+   * @param connector the connector needed for database connection
+   * @param schemaName the database schema to use
+   * @param tableName the database table to use
+   * @param metaInformation the time series meta information
+   * @param timePattern the pattern of time values
+   * @return a SqlTimeSeriesSource for given time series table
+   * @throws SourceException if the column scheme is not supported
+   * @deprecated since 3.0
+   */
+  @Deprecated
+  public static SqlTimeSeriesSource<? extends Value> getSource(
+      SqlConnector connector,
+      String schemaName,
+      String tableName,
+      edu.ie3.datamodel.io.csv.timeseries.IndividualTimeSeriesMetaInformation metaInformation,
+      String timePattern)
+      throws SourceException {
+    if (!TimeSeriesSource.isSchemeAccepted(metaInformation.getColumnScheme()))
+      throw new SourceException(
+          "Unsupported column scheme '" + metaInformation.getColumnScheme() + "'.");
+
+    Class<? extends Value> valClass = metaInformation.getColumnScheme().getValueClass();
+
+    return create(
+        connector, schemaName, tableName, metaInformation.getUuid(), valClass, timePattern);
+  }
 
   /**
    * Factory method to build a source from given meta information