+++ CHANGELOG.md
@@ -26,6 +26,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - BREAKING: Transformer's no load susceptance needs to be zero or negative to pass model validation [#378](https://github.com/ie3-institute/PowerSystemDataModel/issues/378)
   - All input data sets for version < 3.0.0 need to be altered!
+- Deprecating: 
+  - `edu.ie3.datamodel.io.csv.timeseries.ColumnScheme`
+  - `edu.ie3.datamodel.io.csv.FileNameMetaInformation`
+  - `edu.ie3.datamodel.io.csv.timeseries.IndividualTimeSeriesMetaInformation`
+  - `edu.ie3.datamodel.io.csv.timeseries.LoadProfileTimeSeriesMetaInformation`
+  - `edu.ie3.datamodel.io.connectors.CsvFileConnector.CsvIndividualTimeSeriesMetaInformation`
+  - and related methods
 
 ## [2.1.0] - 2022-01-05